Let the demo data provider also stream bytes if the file extension is ".bytes". Change: 139962536
